Feat: [BADA-254] 결제 성공 모달 상태와 로직 관리하는 커스텀 훅 생성 후 적용 #177
